+++ packages/angular-eslint/src/configs/ts-recommended.ts
@@ -25,6 +25,7 @@ export default (
       '@angular-eslint/no-output-rename': 'error',
       '@angular-eslint/no-outputs-metadata-property': 'error',
       '@angular-eslint/prefer-inject': 'error',
+      '@angular-eslint/prefer-on-push-component-change-detection': 'error',
       '@angular-eslint/prefer-standalone': 'error',
       '@angular-eslint/use-pipe-transform-interface': 'error',
       '@angular-eslint/use-lifecycle-interface': 'warn',