+++ packages/eslint-plugin/src/rules/no-async-lifecycle-method.ts
@@ -26,6 +26,7 @@ export default createESLintRule<Options, MessageIds>({
       'Injectable',
       'NgModule',
       'Pipe',
+      'Service',
     ]);
 
     const angularLifecycleMethodsPattern = toPattern([