+++ packages/utils/CHANGELOG.md
@@ -1,3 +1,18 @@
+# 22.0.0 (2026-06-07)
+
+### 🚀 Features
+
+- support the Angular v22 @Service decorator ([#3057](https://github.com/angular-eslint/angular-eslint/pull/3057))
+- ⚠️ drop support for eslint v8 and legacy eslintrc config format ([#3056](https://github.com/angular-eslint/angular-eslint/pull/3056))
+
+### ⚠️ Breaking Changes
+
+- drop support for eslint v8 and legacy eslintrc config format ([#3056](https://github.com/angular-eslint/angular-eslint/pull/3056))
+
+### ❤️ Thank You
+
+- James Henry @JamesHenry
+
 ## 21.4.0 (2026-05-13)
 
 ### 🩹 Fixes